2 Create, import and mount our Vue app
🏳️ Understanding the Vue Instance
When we created our Vue app, we passed in the options object, which allowed us to add some optional properties to configure the application. Doing this creates our Vue instance, the heart of our Vue application, which powers everything.

📄main.js

const app = Vue.createApp({Options Object})
By importing this app, and mounting it to the DOM, we’ve essentially plugged the app into our DOM, giving our HTML a direct line into the app. This way, our template code can access options from the app, such as its data.
If you’re wondering what’s happening with this double curly brace syntax, you can imagine it like a phone, which has access to a phone within our Vue app. From our template, we’re able to ask the app, \“Hey, what’s the value of product?\” And the app responds, \“Socks.\” When the page renders, we see \“Socks\” display on the page.
If this double curly brace syntax, or mustache syntax, is new to you, it allows us to write JavaScript expressions. In other words, it allows us to run valid JavaScript within our HTML.

🏳️ Vue’s Reactivity
What would happen if we changed the value of product from \“Socks\” to \“Boots\”?

📄main.js

const app = Vue.createApp({
    data() {
        return {
            product: 'Boots' // updated data value //
        }
    }
})
Because of how Vue works, the h1's expression that is relying upon product would automatically receive that new value, and our DOM would update to display \“Boots\”.

📄index.html

<div id="app">
  <h1>{{ product }}</h1> <! -- will reactively receive any updates to product -->
</div>
This is because Vue is reactive. Under the hood, Vue has an entire reactivity system that handles updates. When a data value changes, anywhere relying on that data will automatically update for us. We don’t have to do anything to make that happen.
(As a side note, here on Vue Mastery, we have an entire Vue 3 reactivity course. So if you’re interested in exploring the reactivity system under the hood, you can take that advanced course later.)