3.2 Attribute Binding
Create a bond between the src attribute and the data itself using v-bind vue directive. When looking in the browser we see our image displayed.

🏳️ Understanding v-bind
How exactly is the v-bind directive working? We use this directive to dynamically bind an attribute to an expression. In this case, the attribute is src and the expression is whatever is in that attribute’s quotes: "image"

📄index.html

<img v-bind:src="image"> <! -- src attribute bound to the image data -->
If you’re thinking this doesn’t look like a typical JavaScript expression, you can imagine it looking like this: v-bind:src="{{ image }}". Under the hood, Vue is going to evaluate it just the same.

🏳️ A Reactive Bond
Because of Vue’s reactivity system, if we updated our image data to a path that points to the image of our blue socks (image: './assets/images/socks_blue.jpg'), the expression that our src attribute is bound to would update and our browser would display the blue socks image.