4.1 Conditional rendering
Show up the paragraphs depending on if our product is in stock or not.

🏳️ The v-if directive
We can add the v-if directive onto an element to render it based upon a condition, like so:

<p v-if="inStock">In Stock</p>
Now, this element will render only if inStock is truthy.

We can combine the v-if directive with its sister directive v-else to display another element as the fallback if the first condition turns out to be falsey.

📄index.html

<p v-if="inStock">In Stock</p>
<p v-else>Out of Stock</p>
Now, if inStock is false, we’ll see \“Out of Stock\” gets rendered to the page.