4.3 Conditional rendering
Chained Conditional Logic
Earlier, we looked at v-if with v-else, now let’s take a look at how we can add additional layers of conditional logic.

To do this, we’ll replace inStock with inventory:

📄main.js

const app = Vue.createApp({
    data() {
        return {
            ...
            inventory: 100
    }
Since our condition (inventory) is now an integer, we can use a bit more complex logic within our expression. For example:

📄index.html

<p v-if="inventory > 10">In Stock</p>
<p v-else>Out of Stock</p>
Now, we will only render the first p tag if inventory is greater than 10.

Let’s say we want display a new message when the product is almost sold out. In this situation, we could add another conditional level, where we’re watching out for inventory to get below 10 but above 0.

📄index.html

<p v-if="inventory > 10">In Stock</p>
<p v-else>Out of Stock</p>
The v-else-if directive gives us a middle layer of logic. So in this example, if inventory were 8, this p tag would get rendered.
Of course, if inventory is zero, we’ll default to the final level of v-else and display \“Out of stock\”.