5.2. List rendering
🏳️ Product Variant Colors
To get more familiar with list rendering with v-for, we’ll work on another example within our app. Let’s add a new variants array to our data
We now have an array that contains an object for each variant of our product. Each product variant has an id, and a color. So for our next task, we’ll print out each variant color, and use the id to help Vue keep track of our list items.

📄index.html

<div v-for="variant in variants" :key="variant.id">{{ variant.color }}</div>
Notice how we’re using dot notation to print out each variant as we loop through the variants array. But what is that :key attribute doing there?

🏳️ Key Attribute: An essential for list items
By saying :key="variant.id", we’re using the shorthand for v-bind to bind the variant’s id to the key attribute. This gives each DOM element a unique key so that Vue can grasp onto the element and not lose track of it as things update within the app.
This provides some performance improvements, and later down the line, if you’re doing something like animating your elements, you’ll find that the key attribute really helps Vue effectively manage your elements as they move around the DOM.