6.2. Event handling -- Increment the value of cart when 'Add to Cart' button is clicked
🏳️ Listening for Events
In order to know when the button is clicked, we need to be listening for events on that element, specifically click events. We can achieve this by using another Vue directive: v-on.
Here, we are telling v-on what type of event to listen for: a click. Inside the quotes, we place the logic (or method name) we want to run when that event happens.
If we write v-on:click="cart += 1", we’ll increment the value of cart by 1, when a click event happens.