6.3. Event handling
🏳️ Triggering a method
Because the logic cart += 1 is very simple, we could keep it in-line on the button element, like we have it. But often, we need to trigger more complex logic. In those situations, we can add a method name to fire when the event happens. So let’s do that now.
Now, when the button is clicked, the addToCart method will be run. Let’s add that method to our Vue app’s options object.
Notice how we added the methods option, and inside of that we added the new addToCart method, which contains the same logic we just had in-line. The difference here is now we’re saying this.cart to refer to this cart in this Vue instance’s data.
In the browser, we should now be able to click the Add To Cart button and see the value of cart climb up by 1.

🏳️ Understanding v-on
Let’s take a deeper look into how this event handling is working.
By adding v-on to an element, we’re essentially giving it an ear that can listen for events. In this case, we’ve specified that we’re listening for click events. When a click happens, the addToCart method runs, which as we just saw, takes the value of cart and increments it by one.