7.1. Class and style binding
🏳️ Style Binding
In the last lesson, we added the feature where if you hover over “green” or “blue”, you update the image that is being displayed; the green or blue socks, respectively. But wouldn’t the user experience be nicer if instead of hovering over the *word “*green” or “blue”, we hovered over the actual colors green and blue?
Let’s create green and blue circles that we can hover on. We can achieve this by using style binding.
First, to style our divs like circles, we’ll we need to add a new class .color-circle to the variant div.
This class already lives in our css file. As you can see, it simply transform our divs into a circle with a 50px diameter.
Now that we’ve got that out of the way, we can move on to the actual style binding. Just like it sounds, we want to bind styles to the variant divs. We do so by using v-bind (or its shorthand: :) on the style attribute, and binding a style object to it.
Here, we’re setting the divs’ backgroundColor equal to the variant.color. So instead of printing out those strings, “green” and “blue”, we’re using them to set the background color of our circles.
Checking this out in the browser, we should now see two color circles filled in with a green and blue background.
Cool! Now let’s understand, on a deeper level, how this is all working.

🏳️ Understanding Style Binding
On our variant div, we added the style attribute and bound a style object to it.
That style object has the CSS property of backgroundColor, and we’re setting that equal to whatever the variant color is at the time of that v-for iteration.
In the first iteration, variant.color is "green"
Vue takes that information and converts it into the code:style="{ backgroundColor: green }"
Then prints out a green background circle.

🏳️ Camel vs Kebab
There are some important things to consider when using style binding like this.

<div :style="{ backgroundColor: variant.color }"></div>
Inside of this expression, remember that this style object is all JavaScript. That’s why I used camelCase in the property name. If I had said background-color, that - would’ve been interpreted as a minus sign. But we’re not doing any math here. We’re setting a CSS property name.
So since we’re in this JavaScript object, we have to use camelCased unless we want to use ‘kebab-cased’ in quotes to avoid the mathematical misinterpretation, like so:

<div :style="{ 'background-color': variant.color }"></div>
Both options will work, as long as you remember your quotation marks.

🏳️ Style Binding: Objects
Sometimes you might want to add a bunch of styles to an element, but adding them all in-line could get messy. In these situations, we can bind to an entire style object that lives within our data.

📄️ index.html
<div :style="styles"></div>

📄️ main.js
data(){
  return {
    styles: {
      color: 'red',
      fontSize: '14px'
    }
  }
}

Now that we’ve taken a look into the topic of style binding, let’s look at a similar topic: class binding.