7.3. Class and style binding
🏳️ Multiple Class Names
When getting started with class binding, there are some things to note. For example, what happens when we already have an existing class and we want to conditionally add another class based on a data value?
For example, if we already have the color-circle class on this div, and we conditionally add the active class, how will this look?

📄️ index.html
<div class="color-circle" :class="{active: activeClass}"></div>

📄️ main.js
data(){
  return {
    activeClass: true
  }
}

Those classes are going to be combined like so:

<div class="color circle active"></div>

🏳️ Ternary Operators
A helpful tool that class binding gives us is the ability to use in-line ternary operators to add different classes based upon a condition.

📄️ index.html
<div class="color-circle" :class="{isActive ? activeClass : ''}"></div>

📄️ main.js
data(){
  return {
    isActive: true
  }
}