8.1. Computed Properties
🏳️ A Simple Computed Property
In the starting code, you’ll notice we have a new data property. What if we wanted to combine the brand and the product, in our template? We could do that within an expression like so.
If we checked this out in the browser, we’d see “Vue mastery Socks” displayed. But wouldn’t it be neat if, instead of handling this logic in the inner HTML, our app had the ability to compute that value for us? For example, taking the brand and the product, adding them together, and returning that new value.