8.3. Computed Properties -- Update image and inStock using computed properties
🏳️ Computing Image & Quantity
Heading back into our code, let’s add a new quantity property to our variant objects.
Notice how the green socks have a quantity of 50 while the blue socks have 0. In other words, the green socks are in stock and the blue socks are out of stock. However, we’re currently displaying “In stock” or “Out of stock” based on the inStock data value, which no longer reflects the truth about our product and its variant quantities. So we’ll want to create a computed property we can use to display “In stock” or “Out of stock” based on these new quantities.
To get started, remember how we updated the variant image, based on which variant color is moused over? Instead of that mouseover event triggering the updateImage() method, we’re going to have it trigger a new method called updateVariant().
Notice how we’re passing in the index of the currently hovered-on variant: updateVariant(index). We got access to that index by adding it as a second parameter in our v-for directive:

  v-for="(variant, index) in variants"

Why are we passing in the index? We’re going to use it to tell our app which variant is currently hovered on, so it can use that information to trigger the updating of both the image AND whether that variant is in stock or not.
We’ll add a new data property to our app, which will be updated to equal that index
Our updateVariant() method will set the selectedVariant's value equal to the index of the currently hovered-on variant.
Now, we’ve implemented a way for our app to know which product variant is being engaged with, and we’re able to use that information to trigger the computing of which image to show and whether to show “In stock” or “Out of stock”, based on which variant the user is moused over.