8.4. Computed Properties -- Update image and inStock using computed properties
We’re now ready to delete image and inStock from our data, and replace those with computed properties of the same names.
So how do we grab the variant image and quantity? That will look like this:
📄️ main.js

image() {
  return this.variants[this.selectedVariant].image
}
We’re targeting the first or second element of our variants array based off the selectedVariant, which is either 0 or 1, depending on which variant color circle is hovered on. Then we just use dot notation to grab the image off that variant.

The logic for computing inStock is nearly identical:

📄️ main.js

inStock() {
  return this.variants[this.selectedVariant].inStock
}
Checking this out in the browser, when we hover over the color circles, not only are we updating the variant image, but we’re also displaying whether that variant is in stock or out of stock, using its quantity.