9.1. Components & Props
🏳️ Building Blocks of a Vue App
In modern front-end JavaScript frameworks, components are the building blocks of an app, and that’s certainly the case with Vue. You can imagine components a bit like Legos that you can plug into each other in a component family tree hierarchy.
Any given web page may be composed of multiple components, and it’s common for components to be “parents” that have child components nested within them.

🏳️ Creating our First Component
Let’s head in to our app and create our first component. Because our app will eventually have multiple components, we’ll create a components folder, inside of which we’ll create our first component, called ProductDisplay.js.
The syntax to create a component looks like this:

📄️ components/ProductDisplay.js

app.component('product-display', {})
The first argument is the component name, 'product-display' in this case, and the second argument is an object to configure our component (similar to the options object used to configure our root Vue app)

🏳️ Template
Because we need our component to have structure, we’ll add the template property and paste all of the product-based HTML code that currently resides in index.html into here, within a template literal.
Note that we haven’t changed any of this code, we’re simply moving it into the product-display component so that it’s encapsulated there. If you’re wondering what the /*html*/ is doing there, that activates the VS Code extension es6-string-html, which gives us syntax highlighting even though we’re in this template literal.

🏳️ Data & Methods
Now that we’ve given this component its template, or its structure, we need to give it its data and methods, which still live in main.js. So we’ll paste them in now.
We’ll make sure to delete cart from the data here because we don’t need each product to have its own cart.

🏳️ Cleaning up main.js
Now that we’ve encapsulated all of this product-specific code within our product-display component, we can clean up our main.js file.
We’ve left the cart and the methods option because it’ll have a new method later on.

🏳️ Importing the Component
In order to make use of product-display, we need to import it into we’ll index.html. Now that it’s imported, we can use it within our template.
If we check this out in a browser, we’ll see everything is still showing up just like it was before, but now since we’ve rearranged things, the Add to Cart button is no longer incrementing the cart. We’ll fix that in the next lesson.
For now, to show you how helpful these reusable blocks of code can be, I’m going to add two more product-display components.
When we refresh the browser, we’ll see them all showing up. Each of them are independently functional.