9.2. Components & Props
🏳️ Props
Now that we’re starting to learn how to encapsulate reusable code into these components, what happens when our component needs something that is outside of itself? For example, what if the parent, so-to-speak, had some message data, and the child needed it? Because a component has its own isolated scope, it can’t reach up outside of itself to grab something that’s outside of its scope.
The answer here is props. These are custom attributes for passing data into a component. They function kind of like a funnel, into which you can pass the data the component needs.
Let’s add the ability for our product-display component to take in a prop.

🏳️ Giving our component a prop
Let’s give our root Vue app, located in main.js, a new data property, which indicates whether the user was a premium user or not.
If a user is premium, their shipping will be free. So our product-display component needs access to this data. In other words, it needs a custom attribute (a funnel) that we can feed this data into. Let’s add that now, which we’ll do by giving the component a props option, and adding a premium prop to it.
Notice how Vue’s props feature has built-in validation, so we can specify things like the prop’s type and whether it’s required, etc.
Now that we’ve configured this, we can add that custom attribute onto the product-display component where we’re using it.
Notice how we’re using the shorthand for v-bind so we can reactively receive the new value of premium if it updates (from true to false).

🏳️ Using the Prop
Now that our product-display component has the premium prop, we can make use of it within the component. Remember, we want to use the fact a user is premium or not to determine what they need to pay for shipping.
Here, shipping is the name of a new computed property on the product-display component.
The computed property checks whether the premium prop is true, and if so, returns 'Free'. Otherwise, it returns 2.99