10.1. Comunicating events
Our Goal
Give our component the ability to let its parent know an event happened within it.

🏳️ Emitting the event
When we refactored things in our last lesson, moving product-related code into the new product-display component, we broke the ability to click the Add to Cart button and increment the value of cart. This is because the cart lives outside of the product-display component, inside the root Vue app in main.js.
We need to give the product-display component a way to announce that its button is clicked. How do we get this to happen?
We already know that props are a way to pass data down into a component, but what about when something happens within that component, like a button click? How do we let other parts of our app know that that event happened?
The answer is to emit that event, telling the parent that it happened. Let’s add this ability within our product-display component, by amending the addToCart() method.
We’ll write this.$emit() and emit an event called 'add-to-cart'. So when the button is clicked, we’re emitting, or bubbling up, that event. We can listen for that event from within the parent scope, where we’re using product-display, by adding a listener: @add-to-cart.
When that event is “heard” by the parent, it will trigger a new method by the name of updateCart, which we’ll add within main.js.
If we check this out in the browser, we should now be able to click the Add To Cart button, which lets the parent know the add-to-cart event happened, triggering the updateCart() method.