10.2. Comunicating events
🏳️ Adding product id to the cart
To make our app more realistic, our cart shouldn’t just be a number. It should be an array that contains the ids of the products that are added into it. So let’s do a bit of refactoring.
Here, we’ve added a second parameter and grabbed the product’s id much like we grabbed the image and quantity before.
Now in the browser, you can see that we’re adding the product ids into the cart, which is now an array.
But we don’t need to actually display these ids. We just want to display how many items are in the cart. Fortunately, that’s a quick fix.