11.1. Forms & v-model
Our Goal
Create a form for users to add product reviews.

🏳️ Introducing v-model
At the beginning of this course, we learned about v-bind, which creates a one-way binding, from the data to the template. When working with forms, however, this one way binding isn’t enough. We also need to be binding from the template to the data.
For example, when a user inputs their name into an input field, we want to record and store that value in our data. The v-model directive helps us achieve this, creating two-way data binding.
To see this all in action, we’re going to create a new review-form component.

🏳️ The review-form component
We’ll add a new ReviewForm.js file into our components folder, and scaffold out the component.
Inside of our template, notice these elements:

  <input id="name">
  <textarea id="review">
  <select id="rating">