11.2. Forms & v-model
We want to bind the input fields to their respective data properties so that when the user fills out the form, we store their data locally.
We’ll achieve this by adding the v-model directive to each of those input elements.
Notice how on the <select> element, we used v-model.number, this is a modifier that typecasts the value as a number.