11.3. Forms & v-model
🏳️ Submitting the Review Form

In order to submit this form, we’ll add a listener at the top.
We’re using another modifier, @submit.prevent="onSubmit" to prevent the default behavior (a browser refresh). When this form is submitted, it will trigger the onSubmit() method.
That method will create a productReview object, containing the name , review and rating from our data. It will then $emit a review-submitted event, sending that productReview along as the payload.
Finally, we’re clearing out the data fields.