11.4. Forms & v-model
🏳️ Using the Review Form
Now that our review form is created, we can import it within index.html.
Then we’ll head into product-display, and actually use the component within its template, below the “product-container”.
Now in the browser, we can see the review form.
It looks like it’s working… except when we click the submit button, we’re emitting the event, but we haven’t listened for it anywhere. Like we learned in the previous lesson, we need to listen for the review-submitted event in the parent scope (in product-display).
When the event is “heard”, we’ll add the productReview payload to the product-display component’s data.