11.5. Forms & v-model
🏳️ Adding product reviews
We’ll add the event listener here on the review-form, where it’s being used.
When the event happens, we’ll trigger a new addReview() method. This will add product reviews to our product-display component, which means that component needs a new reviews array in its data.
Now let’s flesh out the addReview() method
As you can see, it which takes in the review that we got from the review-submitted event payload, and pushes it into the reviews array.