11.6. Forms & v-model
🏳️ Displaying the reviews
Now that we’ve implemented the ability to add reviews, we need to be displaying those reviews. Let’s create a new component to do that. That component will be called review-list
It will have a prop so that it can receive the reviews and print them out in the template using v-for, including the index, so that we can bind the :key attribute to it.
Now we can import this component inside index.html.
Then add it within product-display, right above the review-form
Notice how we’ve added :reviews="reviews" so we can pass the reviews that live on product-display into the review-list.
Checking this out in the browser, we’ll add a new review, click submit, and see the review is being displayed.
So far so good, but when we refresh (and there are no reviews) we still see an empty box because the review-list component is still being rendered with no reviews to print out.